extracted test class
