+++ RubberduckTests/RubberduckTests.csproj
@@ -123,6 +123,7 @@
     <Compile Include="Parsing\VBACodeStringParserTests.cs" />
     <Compile Include="PreProcessing\CompilationArgumentsCacheTests.cs" />
     <Compile Include="QuickFixes\AccessSheetUsingCodeNameQuickFixTests.cs" />
+    <Compile Include="QuickFixes\IntroduceLocalReDimVariableQuickFixTests.cs" />
     <Compile Include="QuickFixes\IntroduceLocalVariableQuickFixTests.cs" />
     <Compile Include="Rewriter\ArgumentRewriterInfoFinderTests.cs" />
     <Compile Include="Rewriter\ParameterRewriterInfoFinderTests.cs" />
